hindi-tts/requirements.txt
# Core Dependencies
torch>=2.0.0
torchaudio>=2.0.0
chatterbox-tts>=0.1.6

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0
numpy>=1.24.0
audioread>=3.1.0

# ML/DL Frameworks
diffusers>=0.29.0
transformers>=4.30.0
huggingface-hub>=0.36.0
accelerate>=0.20.0

# Utilities
click>=8.3.0
colorama>=0.4.6
joblib>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional - For voice cloning
einops>=0.8.0
conformer>=0.3.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
